Move #getAvroWriter() to AvroUtils object
